fix(Stride): add explicit cast for MipMapLevelOfDetailBias double to float conversion

+++ src/Andastra/Runtime/Graphics/Stride/Graphics/StrideRenderState.cs
@@ -658,11 +658,11 @@ namespace Andastra.Runtime.Stride.Graphics
 
         public float MipMapLevelOfDetailBias
         {
-            get { return _description.MipMapLevelOfDetailBias; }
+            get { return (float)_description.MipMapLevelOfDetailBias; }
             set
             {
                 var desc = _description;
-                desc.MipMapLevelOfDetailBias = value;
+                desc.MipMapLevelOfDetailBias = (double)value;
                 _description = desc;
             }
         }
